# listings/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Listings API
  version: 1.0.0
  description: API for creating and retrieving listings.

servers:
  - url: http://localhost:3000/api
    description: Local dev

tags:
  - name: Listings
    description: Endpoints for creating and managing listings

paths:
  /listings:
    get:
      tags: [Listings]
      summary: List listings
      operationId: listListings
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size
      responses:
        "200":
          description: A page of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingListResponse"
              examples:
                listExample:
                  value:
                    data:
                      - id: "8f8a1c9e-7a2f-4e5b-b6a4-1b2c3d4e5f60"
                        type: "listing"
                        title: "2019 Honda Civic Sport"
                        description: "Well maintained..."
                        location: "Vancouver, BC"
                        endtime: "2025-12-31T23:59:59Z"
                        seller:
                          name: "Alex Doe"
                          location: "Vancouver, BC"
                          id: "user_12345"
                        specifications:
                          engine: "2.0L I4"
                          color: "Blue"
                          model: "Civic Sport"
                    meta:
                      page: 1
                      pageSize: 20
                      total: 1
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Listings]
      summary: Create a new listing
      operationId: createListing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingCreateRequest"
            examples:
              createExample:
                value:
                  data:
                    title: "2019 Honda Civic Sport"
                    description: "Well maintained, single owner, full service history."
                    location: "Vancouver, BC"
                    endtime: "2025-12-31T23:59:59Z"
                    seller:
                      name: "Alex Doe"
                      location: "Vancouver, BC"
                      id: "user_12345"
                    specifications:
                      engine: "2.0L I4"
                      color: "Blue"
                      model: "Civic Sport"
                      transmission: "Manual"
      responses:
        "201":
          description: Listing created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
        "400":
          description: Bad request (malformed JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error (missing/invalid fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /listings/{id}:
    get:
      tags: [Listings]
      summary: Get a listing by ID
      operationId: getListingById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: Listing identifier (UUID or object ID as string)
      responses:
        "200":
          description: A single listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingResponse"
              examples:
                oneExample:
                  value:
                    data:
                      id: "8f8a1c9e-7a2f-4e5b-b6a4-1b2c3d4e5f60"
                      type: "listing"
                      title: "2019 Honda Civic Sport"
                      description: "Well maintained..."
                      location: "Vancouver, BC"
                      endtime: "2025-12-31T23:59:59Z"
                      seller:
                        name: "Alex Doe"
                        location: "Vancouver, BC"
                        id: "user_12345"
                      specifications:
                        engine: "2.0L I4"
                        color: "Blue"
                        model: "Civic Sport"
        "404":
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error: "Not Found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    # ---- Requests ----
    ListingCreateRequest:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ListingCreateBody"

    ListingCreateBody:
      type: object
      required:
        - title
        - description
        - location
        - endtime
        - seller
        - specifications
      properties:
        title:
          type: string
          description: Title of the listing
        description:
          type: string
          description: Long description of the listing
        location:
          type: string
          description: City/Region (free-form for now)
        endtime:
          type: string
          format: date-time
          description: Auction end time in ISO8601
        seller:
          $ref: "#/components/schemas/Seller"
        specifications:
          $ref: "#/components/schemas/Specifications"

    Seller:
      type: object
      required: [name, location, id]
      properties:
        name:
          type: string
        location:
          type: string
        id:
          type: string
          description: Seller identifier (internal/user id)

    # Flexible key/value object; values can be any JSON type
    Specifications:
      type: object
      additionalProperties: true
      description: >
        Open-ended attributes for the item (e.g., engine, color, model).
        Keys are arbitrary strings; values may be any JSON type.

    # ---- Responses ----
    ListingResource:
      type: object
      required:
        - id
        - type
        - title
        - description
        - location
        - endtime
        - seller
        - specifications
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          example: listing
        title:
          type: string
        description:
          type: string
        location:
          type: string
        endtime:
          type: string
          format: date-time
        seller:
          $ref: "#/components/schemas/Seller"
        specifications:
          $ref: "#/components/schemas/Specifications"

    ListingResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ListingResource"

    ListingListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ListingResource"
        meta:
          type: object
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Internal Server Error"

    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "data.title"
              message:
                type: string
                example: "title is required"